+++ outputs-module.tf
@@ -1,5 +1,5 @@
 output "terraform_module" {
-  description = "Information about this Terraform module"
+  description = "Information about this Terraform module."
   value = {
     name       = "bastion-vm"
     version    = file("${path.module}/VERSION")